Define a recommendation for graphql-java-extended-scalars
Pull Request type
----
- [x] Feature

Changes in this PR
----

This commit adds a recommendation for graphql-java-extended-scalars that will accommodate the requirements of compatibly in relationship with graphql-java. The latter is already part of the recommendations.

Per graphql-java-extended-scalars

> use 1.0.1 or above for graphql-java 14.x and above
> use 15.0.0 or above for graphql-java 15.x and above
> use 16.0.0 or above for graphql-java 16.x and above

Reference https://github.com/graphql-java/graphql-java-extended-scalars

The graphql-java-extended-scalars 16.0.0 artifact is not published in Maven Central.
For this reason we are recommending 15.0.0, with a _preferred_ scope, until  https://github.com/graphql-java/graphql-java-extended-scalars/issues/38 is fixed.

Alternatives considered
----

We could not have a recommendation, but this will mean that developers will then have to figure out which version of `graphql-java-extended-scalars` to use according to the version of `graphql-java` that we do recommend.