Fix serialization of List of Enums as Input Arguments
When the target type of an input argument is a list of enumerations we need to transform the
list item to the enum type. Previously the value was left as a string causing unexpected errors.

As part of this change we are moving all the support classes for the  Input Argument serialization tests to:

* If Java, to `com/netflix/graphql/dgs/internal/java/test/` and the name will have a `J` prefix.
  e.g. `JGreetingType`.
* If Kotlin, to `com/netflix/graphql/dgs/internal/kotlin/test/inputobjects.kt` and the name will have a 'K'  prefix.
  e.g. `KGreeingType`