Prevent the creation of Loggers in the Hotpath.
This PR moves all loggers to be available through a `companion object`.
This will prevent instances of loggers to be created as part of the _request execution_/_hotpath_.
Although this was initially identified in the `DataFetcherInvoker`, via a JFR analysis, we are also
favoring consistency and apply the same pattern across all components.