Introduce a CONFLICT error catered to mutations.
Indicates the operation conflicts with the current state of the target resource.
Conflicts are most likely to occur in the context of a mutation.

For example, you may get a CONFLICT when writing a field with a reference value that is
older than the one already on the server, resulting in a version control conflict.

HTTP Mapping: 409 Conflict.
Error Type: FAILED_PRECONDITION