Support Apollo APQ
This commit adds support for Apollos Automated Persisted Queries or APQ.
To enable this feature you will have to set `dgs.graphql.apq.enabled` property to `true`.

You can see a demo of APQ in the following tests:
* SpringMVC with DgsWebMVCAutomatedPersistedQueriesSmokeTest.kt
* WebFlux with DgsWebFluxAutomatedPersistedQueriesSmokeTest.kt

The `PersistedQueryCache` that backs the default  implementation leverages a Caffeine Cache.
You can provide your own `PersistedQueryCache` if the default doesn't suffice,
please review the `PersistedQueryCaffeineCache` if you decide to do so.
It is important that the cache can handle the case where the query text matches the value defined by the
`PersistedQuerySupport.PERSISTED_QUERY_MARKER` property. In the future
we could provide an _adapter_ that facilitates this.