Make fed resolver always apply configured exception handler
Fixes #966 - ensures that a configured exception handler will be applied
to exceptions that occurred within eg a Mono or CompleteableFuture
execution.

The tests specifically for exception handling did not previously verify
that the custom handler was called in all cases; added a check to verify
it was called.

Had to make all of the tests in that suite provide a more fully
configured DFE in order to get them to pass - there was an NPE in
DefaultDataFetcherExceptionHandler when the ExecutionStepInfo wasn't
present.