Upgrading to graphql-java 18.3
Upgrading to graphql-java 18.3 to address a vul available in [18.0, 18.3), more info available at the
[GraphQL Java 18.3 Release Notes].

As part of this we had to pin some dependencies that were causing trouble a and disable dependabot, since its builds
are not reflecting the change they are making properly. In order for dependabot to work we should make sure that
the builds they produce update the dependency locks as well.

[GraphQL Java 18.3 Release Notes]: https://github.com/graphql-java/graphql-java/releases/tag/v18.3