Fix DgsQueryExecutor behavior with respect to HTTP headers
At some point, the various methods in DgsQueryExecutor that accept HttpHeaders were broken,
due to the internal implementation of DGS switching to reading headers from the request.
Work around this by adding a pluggable request customizer, DgsQueryExecutorRequestCustomizer,
which allows for modifying the request inside DgsQueryExecutor. By default, an implementation
of the customizer is installed that will copy any explicitly supplied headers into the request
if it implements MockHttpServletRequest.