Add workaround for newline handling in MultiSourceReader
If one of the schema files being loaded by DgsSchemaProvider ends with
a definition such as a directive, and did not have a newline, the MultiSourceReader
would concatenate the files in such a way that would cause parsing error.

To avoid that, just add a reader that inserts a newline after each schema file.