Fix the inference of the data. path in the response (#1581)
It might be the case that data returned in the data field starts has a
suffix that starts with `data`. This will incorrectly cause the
GraphQLResponse to infer that it shouldn't add the `data.` suffix which
will might cause confusion since the expected behaviour is for the
`data.` suffix to be inferred.