Add fixes from 2.7 (#1584)
* Fix the inference of the data. path in the response (#1581)

It might be the case that data returned in the data field starts has a
suffix that starts with `data`. This will incorrectly cause the
GraphQLResponse to infer that it shouldn't add the `data.` suffix which
will might cause confusion since the expected behaviour is for the
`data.` suffix to be inferred.

* Improve IntelliJ Langauge detection for the graphql string. (#1582)

IntelliJ supports language detection for String templates via the
`@Language` annotation. This will allow the IDE to _color_ the elements
of the template as if they were a GraphQL Query.

---------

Co-authored-by: Bernardo Gomez Palacio <bernardo.gomezpalacio@gmail.com>