Fix breakage when extracting "data" as the path in GraphQLResponse
Recently a change was made in b386856c8 that broke support for extracting
the value from the response using the path "data". While users could simply
use the "dataAsObject" method, there are many instances where users call
extractValue with "data" as the path. Restore support for this, and add
unit tests.