Consistently handle proxied beans in DgsSchemaProvider
Ensure that we are consistently handling proxied beans in DgsSchemaProvider by
refactoring some of the logic to retrieve annotated methods; after fetching
DgsComponent-annotated beans, use a new internal wrapper class, DgsBean, that
correctly handles proxies, and includes a helper method to retrieve all methods
decorated with a particular annotation. In addition to making the handling consistent,
moving the logic to the DgsBean wrapper class avoids repeated unwrapping of proxies
and allocations.