Allow connection directive to be applied to Unions and Interfaces
DgsPaginationTypeDefinitionRegistry generated a connection directive which could only be applied to
Object types; update it to allow Union and Interfaces as well.

There were tests for this scenario but they didn't catch the problem because they never built a GraphQLSchema.
To avoid such issues, switch all of the tests to fully build a GraphQLSchema.