Fallback operation name for anonymous query metrics
Fallback operation name for anonymous queries to be used in metrics tags.

Currently, an unnamed query will show up in metrics as "anonymous". Instead, we generate a better-than-nothing fallback operation name for usage in metrics tags.

In this PR, we could use the selection set to generate a fallback operation name for usage in metrics tags.

Example, the query:

```graphql
{
  foo
}
```

would have the fallback operationName: `-foo`